+++ src/Speaker.ts
@@ -16,10 +16,26 @@ type SpeakerEvents = {
   stop: [];
 };
 
+export type SpeakerPlayOptions = {
+  /**
+   * Whether to interrupt the same track as the `trackId` passed (`track`), all currently playing audio (`all`), or not to interrupt anything (`none`). Default is `none`.
+   */
+  interrupt?: "track" | "all" | "none";
+  /**
+   * If you want to prevent a particular character to speak over themselves, a `trackId` can be set to a unique string. When playing another speech clip, if the same `trackId` is passed and `interrupt` is set to `true`, then the previous clip will stop playing. Default is unset.
+   */
+  trackId?: string;
+};
+
+type SpeakerSource = {
+  sourceNode: AudioBufferSourceNode;
+  trackId?: string;
+};
+
 class Speaker extends EventEmitter<SpeakerEvents> {
   private audioContext: AudioContext | undefined;
 
-  private currentSources: AudioBufferSourceNode[] = [];
+  private currentSources: SpeakerSource[] = [];
 
   public getAudioContext = (): AudioContext => {
     if (this.audioContext) {
@@ -39,8 +55,19 @@ class Speaker extends EventEmitter<SpeakerEvents> {
 
   public play = async (
     audio: ArrayBuffer,
-    interrupt = false,
+    options: boolean | SpeakerPlayOptions = {},
   ): Promise<void> => {
+    // Backwards-compatible with the old boolean `interrupt` parameter
+    if (typeof options === "boolean") {
+      console.warn(
+        "Passing a boolean as the second parameter to `speaker.play()` is deprecated, and should be updated to use an `options` object.",
+      );
+      // eslint-disable-next-line no-param-reassign
+      options = { interrupt: options ? "all" : "none" };
+    }
+
+    const { interrupt = "none", trackId } = options;
+
     const audioContext = this.getAudioContext();
     const source = audioContext.createBufferSource();
     source.connect(audioContext.destination);
@@ -59,19 +86,26 @@ class Speaker extends EventEmitter<SpeakerEvents> {
       source.onended = (): void => {
         resolve();
         this.currentSources = this.currentSources.filter(
-          (currentSource) => currentSource !== source,
+          (currentSource) => currentSource.sourceNode !== source,
         );
         if (this.currentSources.length === 0) {
           this.emit("stop");
         }
       };
-      if (this.currentSources.length > 0 && interrupt) {
-        this.currentSources.map((currentSource) => currentSource.stop());
+      if (this.currentSources.length > 0 && interrupt !== "none") {
+        this.currentSources.forEach((currentSource) => {
+          if (
+            interrupt === "all" ||
+            (interrupt === "track" && currentSource.trackId === trackId)
+          ) {
+            currentSource.sourceNode.stop();
+          }
+        });
       }
       if (this.currentSources.length === 0) {
         this.emit("start");
       }
-      this.currentSources.push(source);
+      this.currentSources.push({ sourceNode: source, trackId });
       source.start();
     });
   };